added atom mapping to vasp and changed 3d view atom properties table so that it has the element symbol which includes the isotope
